+++ lib/commands/generate-examples.ts
@@ -34,6 +34,18 @@ export const builder: yargs.CommandBuilder = {
     describe: "the readme tag name.",
     string: true,
   },
+  max: {
+    alias: "maximumSet",
+    describe: "generate examples by rule MaximumSet.",
+    boolean: true,
+    default: false,
+  },
+  min: {
+    alias: "minimumSet",
+    describe: "generate examples by rule MinimumSet.",
+    boolean: true,
+    default: false,
+  },
 };
 
 export async function handler(argv: yargs.Arguments): Promise<void> {
@@ -44,12 +56,19 @@ export async function handler(argv: yargs.Arguments): Promise<void> {
       consoleLogLevel: argv.logLevel,
       logFilepath: argv.f,
     };
+    let generateRule: "Max" | "Min" | undefined;
+    if (argv.max && argv.min) {
+      generateRule = undefined;
+    } else {
+      generateRule = argv.max ? "Max" : argv.min ? "Min" : undefined;
+    }
     await validate.generateExamples(
       specPath,
       argv.payload,
       argv.o,
       argv.config,
       argv.tag,
+      generateRule,
       vOptions
     );
     return 0;